+++ shell/gpu/gpu_surface_gl.cc
@@ -12,6 +12,8 @@
 #include "third_party/skia/include/core/SkSurface.h"
 #include "third_party/skia/include/gpu/GrBackendSurface.h"
 #include "third_party/skia/include/gpu/GrContextOptions.h"
+#include "third_party/skia/include/gpu/gl/GrGLAssembleInterface.h"
+#include "third_party/skia/include/gpu/gl/GrGLInterface.h"
 
 // These are common defines present on all OpenGL headers. However, we don't
 // want to perform GL header reasolution on each platform we support. So just
@@ -20,6 +22,7 @@
 #define GPU_GL_RGBA8 0x8058
 #define GPU_GL_RGBA4 0x8056
 #define GPU_GL_RGB565 0x8D62
+#define GPU_GL_VERSION 0x1F02
 
 namespace shell {
 
@@ -30,6 +33,9 @@ static const int kGrCacheMaxCount = 8192;
 // cache.
 static const size_t kGrCacheMaxByteSize = 512 * (1 << 20);
 
+// Version string prefix that identifies an OpenGL ES implementation.
+static const char kGLESVersionPrefix[] = "OpenGL ES";
+
 GPUSurfaceGL::GPUSurfaceGL(GPUSurfaceGLDelegate* delegate)
     : delegate_(delegate), weak_factory_(this) {
   if (!delegate_->GLContextMakeCurrent()) {
@@ -37,6 +43,8 @@ GPUSurfaceGL::GPUSurfaceGL(GPUSurfaceGLDelegate* delegate)
         << "Could not make the context current to setup the gr context.";
     return;
   }
+
+  proc_resolver_ = delegate_->GetGLProcResolver();
 
   GrContextOptions options;
 
@@ -52,7 +60,26 @@ GPUSurfaceGL::GPUSurfaceGL(GPUSurfaceGLDelegate* delegate)
   // A similar work-around is also used in shell/common/io_manager.cc.
   options.fDisableGpuYUVConversion = true;
 
-  auto context = GrContext::MakeGL(delegate_->GetGLInterface(), options);
+  sk_sp<const GrGLInterface> interface;
+
+  if (proc_resolver_ == nullptr) {
+    interface = GrGLMakeNativeInterface();
+  } else {
+    auto gl_get_proc = [](void* context,
+                          const char gl_proc_name[]) -> GrGLFuncPtr {
+      return reinterpret_cast<GrGLFuncPtr>(
+          reinterpret_cast<GPUSurfaceGL*>(context)->proc_resolver_(
+              gl_proc_name));
+    };
+
+    if (IsProcResolverOpenGLES()) {
+      interface = GrGLMakeAssembledGLESInterface(this, gl_get_proc);
+    } else {
+      interface = GrGLMakeAssembledGLInterface(this, gl_get_proc);
+    }
+  }
+
+  auto context = GrContext::MakeGL(interface, options);
 
   if (context == nullptr) {
     FML_LOG(ERROR) << "Failed to setup Skia Gr context.";
@@ -78,6 +105,8 @@ GPUSurfaceGL::GPUSurfaceGL(sk_sp<GrContext> gr_context,
     return;
   }
 
+  proc_resolver_ = delegate_->GetGLProcResolver();
+
   delegate_->GLContextClearCurrent();
 
   valid_ = true;
@@ -102,6 +131,20 @@ GPUSurfaceGL::~GPUSurfaceGL() {
   context_ = nullptr;
 
   delegate_->GLContextClearCurrent();
+}
+
+bool GPUSurfaceGL::IsProcResolverOpenGLES() {
+  using GLGetStringProc = const char* (*)(uint32_t);
+  GLGetStringProc gl_get_string =
+      reinterpret_cast<GLGetStringProc>(proc_resolver_("glGetString"));
+  FML_CHECK(gl_get_string)
+      << "The GL proc resolver could not resolve glGetString";
+  const char* gl_version_string = gl_get_string(GPU_GL_VERSION);
+  FML_CHECK(gl_version_string)
+      << "The GL proc resolver's glGetString(GL_VERSION) failed";
+
+  return strncmp(gl_version_string, kGLESVersionPrefix,
+                 strlen(kGLESVersionPrefix)) == 0;
 }
 
 // |shell::Surface|
@@ -327,6 +370,29 @@ flow::ExternalViewEmbedder* GPUSurfaceGL::GetExternalViewEmbedder() {
 // |shell::Surface|
 bool GPUSurfaceGL::MakeRenderContextCurrent() {
   return delegate_->GLContextMakeCurrent();
+}
+
+bool GPUSurfaceGLDelegate::GLContextFBOResetAfterPresent() const {
+  return false;
+}
+
+bool GPUSurfaceGLDelegate::UseOffscreenSurface() const {
+  return false;
+}
+
+SkMatrix GPUSurfaceGLDelegate::GLContextSurfaceTransformation() const {
+  SkMatrix matrix;
+  matrix.setIdentity();
+  return matrix;
+}
+
+flow::ExternalViewEmbedder* GPUSurfaceGLDelegate::GetExternalViewEmbedder() {
+  return nullptr;
+}
+
+GPUSurfaceGLDelegate::GLProcResolver GPUSurfaceGLDelegate::GetGLProcResolver()
+    const {
+  return nullptr;
 }
 
 }  // namespace shell
